+++ jdaviz/configs/mosviz/__init__.py
@@ -1,2 +1,2 @@
 from .plugins import *  # noqa
-from .helper import Mosviz, MosViz  # noqa
+from .helper import Mosviz  # noqa